the phonebook step12 finished

+++ part2/the-phonebook/src/App.js
@@ -4,13 +4,16 @@ import AddPerson from './components/addPerson'
 import People from './components/People'
 import SearchPerson from './components/searchPerson'
 import peopleService from './services/modifyPeople'
+import Notification from './components/Notification'
 
 const App = () => {
 
   const [people, setPeople] = useState([])
-  const [ newName, setNewName ] = useState('')
-  const [ newNumber, setNewNumber ] = useState('')
-  const [ newSearch, setNewSearch ] = useState('')
+  const [newName, setNewName] = useState('')
+  const [newNumber, setNewNumber] = useState('')
+  const [newSearch, setNewSearch] = useState('')
+  const [notification, setNotification] = useState(
+    {message: '', color: '', backgroundColor: ''})
 
   useEffect(() => {
     peopleService
@@ -35,7 +38,7 @@ const App = () => {
       number: newNumber,
     }
 
-    // if person is already added
+    // if person is already added, update number
      if (findPerson !== undefined) {
 
       const updateNumber = window.confirm(
@@ -50,35 +53,61 @@ const App = () => {
               person.id !== findPerson.id ? person : returnedPerson
             )
           )
+          setNotification(
+            {message: `${newPersonObject.name}'s number was updated`,
+            color: 'green', backgroundColor: '#dcedc8'})
         })
         .catch(error => {
-          alert("person number could not be updated")
+          setNotification(
+            {message: `Information of ${newPersonObject.name} has already been removed from server`,
+            color: 'red', backgroundColor: 'mistyrose'})
         })
       }
+    // if person is not added, add new person + number
     } else {
       peopleService
           .create(newPersonObject)
           .then(returnedPerson => {
             setPeople(people.concat(returnedPerson))
+            setNotification(
+              {message: `Added ${newPersonObject.name}`,
+              color: 'green', backgroundColor: '#dcedc8'})
           })
           .catch(error => {
-            alert("Person could not be added to server")
+            setNotification(
+              {message: `${newPersonObject.name} could not be added to server`,
+              color: 'red', backgroundColor: 'mistyrose'})
           })
     }
     setNewName('')
     setNewNumber('')
+
+    setTimeout(() => {
+      setNotification({message: ''})
+    }, 5000)
   }
 
-  const deletePerson = (id) => {
-    peopleService
+  const deletePerson = (id, name) => {
+    const deletePerson = window.confirm(`Delete ${name}?`);
+    if (deletePerson) {
+      peopleService
         .deleteObject(id)
         .then(returnedPerson => {
           setPeople(people.filter(person => person.id !== id))
+          setNotification(
+            {message: `${name} was deleted`,
+            color: 'green', backgroundColor: '#dcedc8'})
         })
         .catch(error => {
-          alert("person could not be deleted")
+          setNotification(
+            {message: `Information of ${name} has already been removed from server`,
+            color: 'red', backgroundColor: 'mistyrose'})
         })
-}
+      setTimeout(() => {
+        setNotification({message: ''})
+      }, 5000)
+    }
+  }
 
   const handlePersonChange = (event) => {
     setNewName(event.target.value)
@@ -95,6 +124,11 @@ const App = () => {
   return (
     <div>
       <h2>Phonebook</h2>
+      <Notification 
+        message={notification.message} 
+        color={notification.color}
+        backgroundColor={notification.backgroundColor}
+        />
       <SearchPerson 
         newSearch={newSearch}
         handleSearchChange={handleSearchChange}
